+++ src/core/State.psm1
@@ -8,7 +8,7 @@
     
 .NOTES
     Module: Win-Debloat7.Core.State
-    Version: 1.2.0
+    Version: 1.2.3
     
 .LINK
     https://learn.microsoft.com/en-us/powershell/scripting/whats-new/what-s-new-in-powershell-75
@@ -30,7 +30,7 @@ class SystemSnapshot {
     [string]$Description
     [hashtable]$Registry
     [array]$Services
-    [string]$Version = "1.2.0"
+    [string]$Version = "1.2.3"
 }
 
 # Cache for version info (PERF-001 fix)