+++ src/core/SystemState.psm1
@@ -8,7 +8,7 @@
     
 .NOTES
     Module: Win-Debloat7.Core.SystemState
-    Version: 1.0.0
+    Version: 1.2.3
 #>
 
 #Requires -Version 7.5
@@ -35,7 +35,8 @@ function Get-WinDebloat7SystemState {
         Telemetry         = (Get-RegistryKey "HKLM:\SOFTWARE\Policies\Microsoft\Windows\DataCollection" "AllowTelemetry") -ne 0
         Location          = (Get-RegistryKey "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\CapabilityAccessManager\ConsentStore\location" "Value") -ne "Deny"
         Copilot           = ((Get-RegistryKey "HKCU:\Software\Policies\Microsoft\Windows\WindowsCopilot" "TurnOffWindowsCopilot") -ne 1) -and ((Get-RegistryKey "HKLM:\SOFTWARE\Policies\Microsoft\Windows\WindowsCopilot" "TurnOffWindowsCopilot") -ne 1)
-        Recall            = (Get-RegistryKey "HKLM:\SOFTWARE\Policies\Microsoft\Windows\WindowsAI" "DisableAIDataAnalysis") -ne 1
+        Recall            = ((Get-RegistryKey "HKLM:\SOFTWARE\Policies\Microsoft\Windows\WindowsAI" "DisableAIDataAnalysis") -ne 1) -or 
+        ((Get-Service "AIFabric*" -ErrorAction SilentlyContinue).Status -eq 'Running')
         
         # Performance / Gaming
         GameMode          = (Get-RegistryKey "HKCU:\Software\Microsoft\GameBar" "AllowAutoGameMode") -ne 0